Fix message header size calculation
The extra byte (to indicate no tagged fields) was introduced in version
2 of the request header. Depending on the versions of client and broker
being used this byte may, or may not, be present. Use the API key and
request API version to determine if the byte should be present in the
request header, rather than assuming it will always be there.

Signed-off-by: Adrian Preston <prestona@uk.ibm.com>