Upgrade to kafka-clients 3.4.0
This is an enabler for integration with the kroxylicious proxy
framework because it depends on kafka-clients too, causing problems
at runtime if we try to call the EncryptionModule built against an
older kafka-clients.

Note: topic-encryption with the vertx-proxy will still requires the
proxied Kafka cluster or client to be less that version 3.1.0
because it does not yet work with topic ids in the FetchResponse.

Signed-off-by: Robert Young <robeyoun@redhat.com>